requirements.txt
ldap3
pyasn1
python-dotenv
dnspython
